requirements.txt
neo4j==6.0.2
openai==2.3.0
pinecone==7.0.0
pyvis==0.3.2
networkx==3.4.2
tqdm==4.67.1
python-dotenv==1.1.1